Update patch set 1
Patch Set 1:

Hi Marios, I think rather that being a swagger file, this generation is really the definition of the "inputs" object. As per https://apihandyman.io/writing-openapi-swagger-specification-tutorial-part-8-splitting-specification-file/ where in the example, the Person definition is placed in a separate, Person.yaml  , we can place the Inputs as a definition in its own Inputs.json file. You see we cannot say this is a swagger or openapi 2.0 specification, as we don't have any paths etc. So maybe just create the Inputs Definition file is the best option ( basically the JSON Schema of the Inputs object ) which is what you have any, so just remove the "title", "swagger" and "defintions" from the JSON generation and you are done

Patch-set: 1